requirements.txt
python-dotenv
numpy 
pandas
setuptools
pymongo 
certifi 
pymongo[srv]==3.12
pymongo
scikit-learn
dill
pyaml
#-e .